estree: fix ${#name} length of native-lifted vars; unblock c_backend test build
010_substring_loop (estree): a top-level string var (s="HELLOWORLD") is
lifted to a native binding (let s = ...), but the ${#s} length lowered
to sh2.getVar("#s") which reads the RUNTIME STORE — empty for a lifted
var — so len=${#s} was 0 and the while loop never ran (empty output).
Emit String(<native read>).length for a  getVar when name is a
lifted or native-store-read-ok binding. Estree gate: 550/551 -> 551/551.

c_backend test: int_render helper put i64 lo/hi into var_ranges which is
(i128, i128) (the range-analysis API migration) — cast, so the lib test
suite compiles. Pre-existing; unrelated to the estree fix.

+++ src/c_backend.rs
@@ -9128,7 +9128,7 @@ mod tests {
     fn int_render(name: &str, lo: i64, hi: i64, w: Width) -> Render {
         let mut r = Render::default();
         r.var_types.insert(name.to_string(), IrType::Int);
-        r.var_ranges.insert(name.to_string(), (lo, hi));
+        r.var_ranges.insert(name.to_string(), (lo as i128, hi as i128));
         r.var_widths.insert(name.to_string(), w);
         r
     }
